test: migrate all unique tests from catch-all into dedicated files
Moves all remaining unique test classes from test_remaining_branches.py
into their respective device/session/helper extended files. Deletes the
empty catch-all. Coverage remains at 100%.
