+++ tests/unit/test_session_auth_extended.py
@@ -353,3 +353,130 @@ class TestHiveRefreshTokensMissingRefreshToken:
         s.auth.refresh_token.return_value = None
         result = await s.hive_refresh_tokens()
         assert result is None
+
+
+# ===========================================================================
+# Migrated from test_remaining_branches.py
+# ===========================================================================
+
+
+class TestRetryWithBackoffNonZeroDelay:
+    """Line 66: asyncio.sleep called when delay > 0."""
+
+    async def test_non_zero_delay_is_awaited_but_succeeds(self):
+        """A non-zero delay entry causes asyncio.sleep to be called; factory still runs."""
+        from unittest.mock import patch
+
+        s = _make_stub()
+        calls = []
+
+        async def factory():
+            calls.append(1)
+            return "ok"
+
+        with patch(
+            "apyhiveapi.session.auth.asyncio.sleep", new_callable=AsyncMock
+        ) as mock_sleep:
+            result = await s._retry_with_backoff(factory, delays=(5,))
+        assert result == "ok"
+        mock_sleep.assert_called_once_with(5)
+        assert len(calls) == 1
+
+    async def test_zero_delay_does_not_call_sleep(self):
+        """A zero delay skips asyncio.sleep."""
+        from unittest.mock import patch
+
+        s = _make_stub()
+
+        async def factory():
+            return "done"
+
+        with patch(
+            "apyhiveapi.session.auth.asyncio.sleep", new_callable=AsyncMock
+        ) as mock_sleep:
+            result = await s._retry_with_backoff(factory, delays=(0,))
+        assert result == "done"
+        mock_sleep.assert_not_called()
+
+
+class TestUpdateTokensFlatDictWithExpiresIn:
+    """Lines 100->106: flat token dict with ExpiresIn sets token_expiry."""
+
+    async def test_flat_dict_with_expires_in_sets_token_expiry(self):
+        """Flat token dict containing ExpiresIn updates tokens.token_expiry."""
+        s = _make_stub()
+        flat = {
+            "token": "t",
+            "refreshToken": "r",
+            "accessToken": "a",
+            "ExpiresIn": 1800,
+        }
+        await s.update_tokens(flat)
+        assert s.tokens.token_expiry == timedelta(seconds=1800)
+
+    async def test_flat_dict_tokens_are_stored(self):
+        """All token values from flat dict are written to token_data."""
+        s = _make_stub()
+        flat = {"token": "my-id", "refreshToken": "my-rt", "accessToken": "my-at"}
+        await s.update_tokens(flat)
+        assert s.tokens.token_data["token"] == "my-id"
+        assert s.tokens.token_data["refreshToken"] == "my-rt"
+        assert s.tokens.token_data["accessToken"] == "my-at"
+
+
+class TestLoginApiError:
+    """Lines 160-162: HiveApiError in login() is logged and re-raised."""
+
+    async def test_login_api_error_reraises(self):
+        """HiveApiError raised by auth.login propagates unchanged to the caller."""
+        s = _make_stub()
+        s.auth.login.side_effect = HiveApiError()
+        with pytest.raises(HiveApiError):
+            await s.login()
+
+
+class TestHiveRefreshTokensNoAuthResult:
+    """Lines 341->373: refresh returns a result but without AuthenticationResult."""
+
+    async def test_result_without_auth_result_does_not_update_tokens(self):
+        """When refresh_token returns a dict with no AuthenticationResult, tokens stay unchanged."""
+        s = _make_stub()
+        s.tokens.token_created = datetime.now() - timedelta(hours=2)
+        s.tokens.token_expiry = timedelta(hours=1)
+        # Return something truthy but without AuthenticationResult
+        s.auth.refresh_token.return_value = {"SomeOtherKey": "value"}
+        result = await s.hive_refresh_tokens()
+        # Tokens must not have been updated
+        assert s.tokens.token_data["token"] == ""
+        assert s.tokens.token_data["accessToken"] == ""
+        # result is what refresh_token returned
+        assert result == {"SomeOtherKey": "value"}
+
+    async def test_none_refresh_result_does_not_update_tokens(self):
+        """When refresh_token returns None, tokens are left unchanged."""
+        s = _make_stub()
+        s.tokens.token_created = datetime.now() - timedelta(hours=2)
+        s.tokens.token_expiry = timedelta(hours=1)
+        s.auth.refresh_token.return_value = None
+        await s.hive_refresh_tokens()
+        assert s.tokens.token_data["token"] == ""
+
+
+class TestUpdateTokensUnknownKey:
+    """session/auth.py 100->106: tokens dict has neither AuthenticationResult nor token."""
+
+    async def test_unknown_key_does_not_raise_and_does_not_update_tokens(self):
+        """When neither expected key is present, data stays {}, ExpiresIn check skips."""
+        s = _make_stub()
+        original_token = s.tokens.token_data["token"]
+        # Pass a dict that is neither the AuthResult form nor the flat-token form
+        await s.update_tokens({"some_other_key": "some_value"})
+        # Tokens must be unchanged
+        assert s.tokens.token_data["token"] == original_token
+
+    async def test_unknown_key_does_not_set_token_expiry(self):
+        """ExpiresIn check at line 106 skips when data is {} (no match in either branch)."""
+        s = _make_stub()
+        original_expiry = s.tokens.token_expiry
+        await s.update_tokens({"random_key": "random_value"})
+        assert s.tokens.token_expiry == original_expiry
